+++ integration_tests/suite/helpers/helpers/__init__.py
@@ -89,6 +89,7 @@ from . import destination
 
 from . import agent
 from . import agent_login_status
+from . import agent_skill
 from . import call_filter
 from . import call_filter_entity
 from . import call_filter_recipient_user
@@ -182,6 +183,7 @@ __all__ = [
 
     'agent',
     'agent_login_status',
+    'agent_skill',
     'call_filter',
     'call_filter_entity',
     'call_filter_recipient_user',